Set default thread count for Intel GPUs to 32
Changed the default thread count for Intel GPUs (both dGPU and iGPU)
to 32. This is likely because only newer iGPUs are supported by the
Intel Compute Runtime (2024+), and dGPUs already prefer 64
threads (2x32).

As a result, removed the module_unstable_warning() entry for Intel
GPUs on non-Apple OpenCL platform for the following hash modes:
8200, 17200, 17220, 17225, 21700, 25000, 25100, and 25200.