Fix RC4 primitives bug on GPUs with thread size != 32
Fixed a bug in the RC4 crypto primitives that can cause false
negatives on GPUs which do not result in a thread size of 32 in
module_jit_build_options(). This typically affects Intel GPUs
since those often have a native thread size of 8, but not NVIDIA
or AMD.

The problem was in the key lookup logic in inc_cipher_rc4.cl. It was
written in a way that assumes 32 threads are always running:

If the kernel is launched with only 8 threads, this macro can result
in out-of-bound reads and writes.

To fix this, we now enforce that all kernels using the RC4 primitives
are launched with 32 threads (on GPUs).

The following hash modes were updated to include this fix:

7500, 9700, 9710, 9720, 9800, 9810, 9820, 10400, 10410, 10420, 10500,
10510, 13100, 18200, 25400, 33500, 33501, and 33502.