Update method to query free memory on Linux to /proc/meminfo
Previously we used sysinfo(), but when large amounts of memory are
used by the filesystem cache, which still counts as potentially free,
this method fails. We now query /proc/meminfo, if it exists, and
check the MemAvailable attribute. Only if the file is missing do we
fall back to sysinfo().