Warn instead of waiting on high GPU memory usage
For algorithms using module_extra_tuningdb_block(), such as scrypt,
Argon2, or LUKS2, we changed the behavior when a compute device shows
high memory usage. This typically comes from the desktop window manager,
desktop applications, or other GPGPU programs using GPU acceleration
and occupying memory.

In v7.0.0, hashcat printed a warning 10 times with delays between each,
allowing the user to close applications and free memory. On systems with
multiple GPUs, this could lead to significant wait times if memory was
not freed.

We now show a single warning instead of waiting. The user is informed
about the expected performance loss in percent. To avoid spamming and
false positives, the warning is shown only if memory usage exceeds a
default threshold of 15 percent.