Add argon2_init() wrappers to support private address space
In release version v7.0.0, argon2_init() enforced the use of buffers
from the GLOBAL_AS address space for both the password and the salt.
This requirement cannot always be fulfilled by kernel developers,
because future kernels may not provide the password as is.

For example, an application may use sha256(password) and pass the
result to argon2_init() instead.

Salt preprocessing may also be needed. For instance, some formats
require byte swapping or other transformations before use.

To support these cases, we now provide appropriate helper functions
for kernel developers. This is done in the same way as with scrypt,
by creating wrapper functions that handle address space conversion.
This makes usage simpler for plugin developers.

The four new functions are:

  argon2_init_pp (PRIVATE_AS const pw_t *pw, PRIVATE_AS const salt_t*salt, ...
  argon2_init_pg (PRIVATE_AS const pw_t *pw, GLOBAL_AS  const salt_t*salt, ...
  argon2_init_gp (GLOBAL_AS  const pw_t *pw, PRIVATE_AS const salt_t*salt, ...
  argon2_init_gg (GLOBAL_AS  const pw_t *pw, GLOBAL_AS  const salt_t*salt, ...

The letter p stands for PRIVATE_AS and g stands for GLOBAL_AS.

As an example, in hash-mode 34000, the following change was required:

  argon2_init (&pws[gid], &salt_bufs[SALT_POS_HOST], &options, argon2_block);

to:

  argon2_init_gg (&pws[gid], &salt_bufs[SALT_POS_HOST], &options, argon2_block);

That is all.