Fix python bridge early problems
1. Fix unsalted hashlist support

The python bridge was designed for salted and slow hashes. For
that reason it is implemented as an outside mode (means the
password candidate generator sits outside the most deep guessing
nested loop). In outside mode, special treatment is required to
allow multihashing. This is typically done through the flag
OPTS_TYPE_DEEP_COMP_KERNEL in the module.

This flag, when used, requires the plugin developer to specify a
specific kernel instead of the typical mXXXXX_comp kernel. See
the plugin developer documentation at
docs/hashcat-plugin-development-guide.md for how to use this
attribute.

There is a lesser known and currently undocumented alternative
called OPTS_TYPE_MULTIHASH_DESPITE_ESALT which allows a plugin
developer to ignore hashcat internal checking routines that
prevent it from running when an outside kernel is detected with
multiple salts and multiple esalts, for which
OPTS_TYPE_DEEP_COMP_KERNEL is usually required. To use it, the
module developer must ensure a one to one relationship between
salts and esalts. That is typically done in the module decoder
sections where values from esalt are used to generate a unique
value for salt, for example by creating a hash from the esalt
data and storing it in the salt field.

The OPTS_TYPE_MULTIHASH_DESPITE_ESALT has been enabled to allow
the python bridge to be used efficiently with unsalted hashes.

2. The esalt structure was too large

For most use cases that is irrelevant but in cases with 200000
hashes to be cracked the 128KiB buffer is a multiplier ending up
with a 24GiB buffer requirement just for the esalt data.

typedef struct
{
  u32 hash_buf[16384];
  u32 hash_len;

  u32 salt_buf[16384];
  u32 salt_len;
} generic_io_t;

This was simply overdesigned because the use case was too
specific. The original idea was that developers would deal with
large buffers, for instance crypto texts, and that it would be
necessary to store some data allowing for decryption in the
python code.

I have reduced that buffer to a much smaller size of 1KiB. If
more memory per hash is needed in the future this is the
structure to increase. Keep in mind there is a copy of this
structure in the JIT compiler world so you must synchronize them.

typedef struct
{
  u32 hash_buf[256];
  u32 hash_len;

  u32 salt_buf[256];
  u32 salt_len;
} generic_io_t;

3. The original idea of the python bridge

The original workflow produces one hash per password and the idea
was that the developer returns one test result (of a successful
decrypt) or hash in its final form (not binary).

This is a key concept in the python bridge because we want to
enable the user to avoid writing a parser for the hashes. The
goal is to enable rapid prototyping. If the developer must write
a decoder and an encoder then the whole idea is lost.

But in hashcat, there must be a decoder and an encoder and they
exist. The decoder and encoder in generic mode work
transparently by hashing the user given hash into an MD4 hash
which is then stored as the target hash in the hashcat database.
Later in the comparison stage we do the same. We take the
returned value from the user, hash it with MD4, and find it in
our database.

The python bridge takes the password from pw_buf[], copies it to
the calc_hash() function we write, and what we return is then
copied to out_buf[] and the MD4 is done on it on the GPU in the
_comp kernel.

typedef struct
{
  u32 pw_buf[64];
  u32 pw_len;

  u32 out_buf[64];
  u32 out_len;
} generic_io_tmp_t;

We improved this to a one to many rate.

typedef struct
{
  u32 pw_buf[64];
  u32 pw_len;

  u32 out_buf[32][64];
  u32 out_len[32];
  u32 out_cnt;
} generic_io_tmp_t;

Now there is space for 32 return hashes. The python bridge
endpoint will check the return value type first. If it is a
string it will copy it to the first element, preserving backward
compatibility. If it detects a list it will iterate over the
list. For one to many hashes return a list instead.